Port over an end-to-end test with complete logic to in-cluster test
